pyproject.toml
[tool.poetry]
name = "test123"
version = "0.1.0"
description = ""
authors = ["This is a test"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
loguru = "^0.7.0"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"